+++ Library/Aphid/Model.js
@@ -287,6 +287,14 @@ Aphid.Model = Class.create({
   **/
   isLoaded: false,
 
+  /**
+   * Aphid.Model#errors -> Array | false
+   *
+   * An array that is populated with any error messages after failing a call
+   * to [[Aphid.Model#validate]], or false if there are no validation errors.
+  **/
+  errors: false,
+
   // -------------------------------------------------------------------------
 
   /**
@@ -577,6 +585,23 @@ Aphid.Model = Class.create({
   {
     $L.info("Saving...", this.displayName);
 
+    // Validate
+    this.errors = false;
+    if (!this.validate())
+    {
+      var errorMessage = new Element("p").update("Errors are present that are preventing your changes from being saved: ");
+      var errorList = new Element("ul");
+      errorList.insert(this.errors.invoke("toElement"));
+      errorMessage.insert(errorList);
+
+      var alertView = new Aphid.UI.AlertView();
+      alertView.title = "Unable to Save";
+      alertView.message = errorMessage;
+      alertView.status = this.displayName;
+      alertView.showAnimated();
+      return false;
+    }
+
     // Assemble URL
     var urlTemplate = new Template(this.url);
     // TODO Make the identifier field configurable
@@ -645,6 +670,43 @@ Aphid.Model = Class.create({
     new Ajax.Request(url, options);
   },
 
+  // Validation --------------------------------------------------------------
+
+  validate: function()
+  {
+    this.errors = $A();
+
+    $H(this.proxies).keys().each(
+      function(proxyAttribute)
+      {
+        if (Object.isArray(this[proxyAttribute]))
+          this[proxyAttribute].each(
+            function(instance)
+            {
+              instance.validate();
+              this.errors.push(instance.errors);
+            }.bind(this)
+          );
+        else
+        {
+          this[proxyAttribute].validate();
+          this.errors.push(this[proxyAttribute].errors);
+        }
+      }.bind(this)
+    );
+
+    // Combine Errors
+    this.errors = this.errors.flatten();
+
+    return (this.errors.length == 0);
+  },
+
+  addError: function(message, field)
+  {
+    if (!this.errors) this.errors = $A();
+    this.errors.push(new Aphid.Model.Error(message, field));
+  },
+
   // Callbacks ---------------------------------------------------------------
 
   _afterLoad: function()
@@ -692,3 +754,5 @@ Aphid.Model = Class.create({
   }
 
 });
+
+//= require "Model/Error"